refactor(patch-detail): extract code into reusable sub-components
